// Backend/apps/contracts/typechain-types/EvidenceRegistry.ts
export declare namespace EvidenceRegistry {
  export type EvidenceStruct = {
    caseId: string;
    version: BigNumberish;
    reportHash: BytesLike;
    timestamp: BigNumberish;
    submittedBy: AddressLike;
  };

  export type EvidenceStructOutput = [
    caseId: string,
    version: bigint,
    reportHash: string,
    timestamp: bigint,
    submittedBy: string
  ] & {
    caseId: string;
    version: bigint;
    reportHash: string;
    timestamp: bigint;
    submittedBy: string;
  };
}